Poke remote CPUs when many elements in RCU list
If we have systems with a bunch of idle CPUs and few busy CPUs where
lot of kernel resources are scheduled to be freed - we may run into the
situation that the rcu list gets not freed up quick enough on the busy CPUs.
The reason is that the idle CPUs wake up once every second and call
the RCU::update().

The commit adds the feature to send an explicit IPI to the other threads when
a number of elements in the rcu queues exceeds a specific limit. On the
receiver CPUs RCU::update get invoked iif the CPUs is/was in idle state.

It's a workaround - not a fix. :-(

What is the right limit ?
-------------------------
Limit to low  -> much IPI overhead, performance penalty
Limit to high -> higher probability to trigger "out of memory" of kernel