Workaround for non-invariant TSCs
On some older CPUs, in particular:

"6:17:a:0 [a07] Intel(R) Core(TM)2 Duo CPU     E8400  @ 3.00GHz"

the TSC can stop in CPU sleep states (non-invariant TSC). This can lead to
the situation that an LAPIC timer interrupt occurs (after the programmed
number of bus clock ticks), but the current TSC is lower than the next
timeout value in the 'Timeout' list. In that case, the LAPIC timer will
not be reprogrammed for the next timeout, because currently that happens
only when a 'Timeout' object gets removed from (or inserted into) the
list.

This workaround reprograms the LAPIC timer in the case that no 'Timeout'
object was removed from the list.