Let the VMM update the PDPTE registers if needed
According to the Intel manual section 4.4.1 of Vol. 3A, the PDPTE
registers in the CPU get updated as part of certain 'MOV to CR'
operations. It seems that the registers need to get updated manually
if the operation caused a VM exit and nested paging is enabled,
because according to the Intel manual section 26.3.2.4 of Vol. 3C,
on the following VM entry the values of the registers are loaded
from the VMCS structure.

This patch lets the NOVA kernel call the 'control register access' VM
exit handler of the VMM, if necessary, to let it update the PDPTE
registers with the values stored in guest-physical memory.