Avoid writing to boot commandline of hypervisor
NOVA checks all passed arguments in the command line to figure out
which of the supported arguments are set. This has previously been
done by splitting the complete string (containing all arguments) into
several strings by replacing the argument separators with '\0'. This
commit changes the argument detection and removes the overwrite.

At the same time, this commit prepares multiboot2. Multiboot1 provides
a pointer to the kernel command line in the multiboot information
structure. This is a physical address and needs to be mapped. With
multiboot2, the string is directly embedded in the multiboot
information which is already mapped. Hence, the mapping of the
command line changes in order to pave the way for multiboot2.