multiboot: fix booting with Qemu's multiboot loader
Qemu supports booting Multiboot kernels via its -kernel option. This
currently doesn't work, because the multiboot header is not in the
first 8K of the produced binary. This seems to be a regression
introduced by building with Tup, because the original Makefile makes
sure to link the assembly files first and thus avoids this problem.

Move the Multiboot entry point, which also contains the Multiboot
header, into its own section and link it at the beginning of the
binary to avoid this.

With this change Qemu is able to boot NOVA directly again.