fpu: don't leak FPU state into newly created ECs
FPU state is not allocated when an EC is created. Instead, it is
created lazily when NOVA has to save FPU state for an EC for the first
time. When an EC touches FPU registers for the first time there is no
FPU state to load and NOVA sanitizes the current state using FNINIT.

FNINIT cleans neither legacy/MMX registers nor SSE registers. Using
only FNINIT to sanitize FPU state causes content left in these
registers from previous ECs to be available to newly created ECs. This
is an info leak vulnerability.

Allocate FPU state on creation instead to avoid this situation. Also
mask SSE exceptions in MXCSR as demanded by the System V ABI.