fpu: expose XSAVE to guests
By allowing the VMM and guests to modify XCR0, we effectively allow
guests to use modern instruction set extensions as well.

XCR0 is now unconditionally saved and restored on VM exits, even for
exits that the kernel handles and where we could leave the guest value
in place. Once we have microbenchmarks in place, we can see whether
there is any merit in optimizing this.

The VM entry path will check whether XCR0 is correct and will kill the
vCPU, in case it is not.

Fixes #18.