tss: remove mentions of tss_handler and the dbf TSS
There is no task switching in 64-bit mode, so using task gates for
double faults and other exceptions that need a fresh stack is not
helpful. If these task gates are invoked, another #GP is triggered and
we end up in a triple fault.

So remove the dbf TSS for now. Because the selector in the task gate
doesn't matter, we can just as well use our normal task gate selector.