tss: don't use cpulocal memory for TSS segments
TSS segments are annoying, because they have to be located within 64K
of the IO permission bitmap. This makes it infeasible to just randomly
allocate them anywhere. The previous solution used CPU-local memory to
achieve closeness between TSS and IO permission bitmap.

As we want to get rid of CPU-local memory, place a special virtual
memory region right in front of the IO permission bitmap and use it
solely for the TSS of each CPU. This region is not CPU-local anymore.