ec: remove hardcoded Exc_regs size from assembly
We use a pointer to Exc_regs for all entry paths. In the system call
entry path, we actually need a pointer to Sys_regs and manually
correct for the offset. This correction hardcodes the size differences
between those structs.

Instead, just store the actual address of where we want to dump user
state.

Suggested-by: Markus Napierkowski <markus.napierkowski@cyberus-technology.de>