x86: rename flush to clflush
There are quite a few functions in the code with the name "flush" that
do different things. Especially in the page table code, flush can
either be a cache flush or a TLB flush resulting in a hard time
reading the code. Fix this by renaming cache flushes to the easily
recognizable clflush.