pte: don't encode mapping order in EPT entries for no reason
The page table modification code had support for embedding the size of
mappings into the page table entries. This was used in the EPT code
only to return the original size that was mapped on not the size of a
single page table entry for a call to lookup().

As there is actually no call to Ept::lookup() in the kernel there is
no reason to special case the behavior of lookup(). Remove this code
instead.