pte: sanity check page table walking
The Pte::walk() function never checks that page table entries are
actually normal entries and not super pages. This is a scenario that
is currently prevented by the way the mapping database works. As this
is a pretty large footgun, assert correct use in the code, so we don't
accidentally break it later.