cpu: clear RFLAGS.NT and RFLAGS.TF on system call entry
For x86_64, NOVA never cleared the trap and nested task flags. Both
can be used to crash the kernel:

Setting the trap flag when entering the kernel breaks immediately,
because the CPU generates a #DB exception on the first kernel
instruction that NOVA cannot handle. This is especially bad, because
this exception will be caught on the user's stack pointer. Without
SMAP, this might lead to arbitrary code execution in the kernel.

Setting the nested task flag crashes the kernel, when entering the
kernel via syscall and exiting via IRET. This happens for example for
a blocking semaphore down operation.