build: enable -ffreestanding
When including headers for unit tests, the headers need to know
whether they are being used in a hosted or bare-metal environment. The
cleanest solution I could come up with is to use -ffreestanding, which
among other things, causes the macro __STDC_HOSTED__ to be undefined.

Use this feature to make assert.hpp compatible to being used in our
unit tests.