hpt: enable NX support
NOVA currently sets HPT_NX to zero, preventing the creation of
no-execute memory both in userspace as well as in the kernel. As this
is an important security features, let's enable it egain.

This patch enables NX support in the IA32_EFER MSR early during boot.
This means that systems that do not support NX will not boot anymore.
To my knowledge NX is available on all 64-bit x86 systems, so there
shouldn't be a problem.

Userspace HPT mappings now respect the execute permission bit. The
initial kernel page table that is set up in assembly does not set NX
bits yet, but all mappings that are done afterwards in the kernel
properly set NX bits.