pd: don't use mdb for memory mappings anymore
Make Space_mem independent from Space and thus remove the ability
to store memory mapping information in the mapping database. The
casualty is revoke support for memory, which is now only
rudimentary and can only revoke everything. The intention is to
remove revoke support completely once userspace is happy with it.