test: don't auto-discover catch2 tests
When messed up linking in the unit test and the resulting binary was
not able to print the list of tests it contained, this resulted in no
unit tests being run. Unfortunately, ctest would still count this as
success.

So instead of living with this risk, we manually add the unit test
binary as a single test and live without the fancy output.