lapic: do not clobber low memory on AP bringup
During AP startup, we need the AP startup code to live in low memory,
because this is how the bootup sequence works. So far, we just
overwrite memory at address 0x1000 and hoped that no one noticed.

It turns out that this strategy doesn't work on UEFI, especially if
NOVA is booted as part of the ExitBootServices hook, where all memory
is potentially in use.

To deal with this, we copy the AP bootup code as part of the
INIT-SIPI-SIPI sequence and restore the original memory once all CPUs
have arrived.