ec: allow user to select in which PD UTCB / vLAPIC page is mapped
Until now, the UTCB / vLAPIC page was always mapped into the owning PD
of an EC, i.e. the PD in which the EC executed. This is inconvenient,
if for example vCPUs are created in otherwise empty PDs.

To make this less painful allow the user to select whether the UTCB /
vLAPIC page is mapped into the owning PD or the calling PD.