kobj: add special purpose capability_cast operator
Multiple places replicate the logic of how to check whether a
capability can be safely converted into a KObject pointer. Also
"sanitize_cap" is ill-named, because it doesn't say what the actual
intention is. Finally, all uses of sanitize_cap are error prone,
because checking for the type marker and static_casting to the desired
place happen in different places and allow for the types to be messed
up.

Instead, introduce a capability_cast operator that safely converts
capabilities to the correct pointer type.