build: fix linking with clang in debug mode
Clang 9 discards a section that is still referenced from our .fixup
section. This results in a broken build. This seems to be related with
this section being emitted from a template instantiation. As the
templating in the MSR code is somewhat pointless anyhow, get rid of it
and thus restore the clang build.