ioapic: add support for state saving and restoring
Change the IOAPIC to include a copy of its redirection table. This
means that it is now large and having a slab allocator for it, makes
no sense. Hence we allocate one whole page for each IOAPIC.

On suspend, we dump all redirection table entries and restore them 1:1
on resume.