hip: mark INIT memory as reserved
For resume, we need to reuse the initialization code. This code used
to be discarded after boot by marking it with INIT and linking it into
a special section. This memory was then not covered by the special
memory region that describes the hypervisor memory.

Change the microhypervisor section of the HIP to include all
hypervisor code.

While I was here, I reduced the alignment for the rest of the
hypervisor code to 2 MiB from 4 MiB to waste less memory for
userspace.

Debugged-by: Markus Partheymueller <markus.partheymueller@cyberus-technology.de>