acpi: reinitialize after suspend and S1 support
For S1, we need to be able to jump into our initialization code after
programming the sleep state. This works by spinning on the WAK_STS bit
in PM1_STS. Since this bit will never be set for S2 and S3 sleep
states, we can use the same code path for those as well. This is
convenient, because we have no way to distinguish the desired sleep
states without a DSDT parser.

Also re-initialize ACPI after resume, because it is not guaranteed
that we come up with SCI_EN set.