acpi: remove ACPI interrupt (SCI) handling
We used to program the ACPI timer overflow interrupt, but we didn't do
anything with the information. We also enabled other ACPI events
without consuming them.

To avoid messing with userspace's configuration of these settings,
remove this behavior and don't handle the SCI in NOVA at all, but pass
it as a normal interrupt to userspace to deal with.