boot: move boot stack into high memory
The boot stack is created in the linker script and is covered by
neither ELF segment. I'm not sure why this happened, but it leaves a
footgun lying around, because we have to remember that there is memory
that is being used by Hedron that is special in that sense.