memory: move most code into high memory
Since suspend/resume support, there is no reason to link
initialization code into a special section to discard it. We need this
code to stay around. This means we can also get rid of the special
markers for these functions.

This is the first step that only makes the relevant attributes a
no-op.