hip: expose two hypervisor memory regions
When we changed marked the INIT code and data as being used by the
hypervisor, we created a situation where the HIP will claim all
physical memory between LOAD_ADDR and LINK_E to be used.

This clashes with our UEFI loader that will only mark the ELF segments
as used, which leaves a usable gap in the memory map that NOVA will
not hand out.