pd: fix golden hpt byte/page confusion
The hypervisor keeps a "golden" page table that is used to remember
all physical memory that the hypervisor is willing to give to
userspace. This is constructed in Pd::Pd(), the constructor for
Pd::kern.

In this code, we confused bytes and pages and thus pretty much
everything was mappable by userspace. Fix this by introducing a small
helper function that does the conversion.