vmx: don't die on VM entry failures
Userspace can easily construct UTCB states that result in VM entry
failures. As VM entry failures result in the kernel crashing, this is
an easy DoS. This means we need to avoid this scenario.

Instead of checking the guest state ourselves, we let the CPU do it
and in case VMRESUME/VMLAUNCH fails, we just consider this an invalid
state exit and traverse through the same portal (0xfd) as the SVM code
when it encounters this condition.