vmx: allocate vpids from the correct CPU
VPIDs need to be unique on the CPU where they are used, but until now
we allocated them on the CPU where they are created with create_ec.
This is broken, because it can lead to vCPUs with the same VPID
running on the same CPU, which in turn messes up their view of virtual
memory.

Change the code to use the correct VPID counter to allocate the VPIDs
so this cannot happen anymore.