ci: workaround pexpect errors
Sometimes we see the following error in Hydra when trying to terminate
a qemu instance that runs the integration test:

pexpect.exceptions.ExceptionPexpect: Could not terminate the child

This seems to be caused by pexpect having an extremely short timeout
when killing processes. Increase the timeout and hope that this
improves stability of the test.

See this issue for details:
https://github.com/pexpect/pexpect/issues/462