pci: fix regression in claim_dev
While refactoring this function to use algorithms, The original code
skipped the first device when looking for deeper PCI devices. The new
code does not. Because the current device is never deeper in the
hierarchy than itself, this results in no devices to be claimed.

Fix the code to skip the initial PCI device when looking for more
devices to claim.

Fixes: 00bb75c41cfe45b7e4d4e6325fda92fc1b629070 ("pci: use list iterators and algorithms")