cpu: Enable FSGSBASE
This is a required feature now. There is a panic if Hedron is booted on
machines that do not support FSGSBASE. The feature is required now
because we want to support Thread Local Storage and the read and write
fsgsbase instructions make it convenient and efficient to handle fs/gs base in the
userspace.