fsgs: use rd/wr fs/gs base instead of rd/wr MSR
We see a heavy performance regression when using read and write MSR to
context switch fs and gs for ECs. We switch to use rdfs/gsbase and
wrfs/gsbase as those instruction should be way faster. This even
compensates for the double swapgs that is now necessary, because the
kernel calls swapgs on kernel entry and exit.