pd: use a TLB invalidation counter instead of counting all IPIs
Simplify and explain the existing TLB invalidation logic by removing
the IPI counters and replacing them by a single integer that counts
TLB invalidations.

The only functional change here is that we increase the TLB counter
early to avoid wasting time in the shootdown loop.

I've also changed the increase operation to an atomic add, because it
is concurrently written and read. Not using atomics on both sides
feels like violating the C++ data race rules.

The Atomic::add is probably more synchronization than we need, because
it results in a sequentially consistent add, while we only need
relaxed atomics. As long as it is not a performance issue, I'd rather
err on the side of caution here.