Initialize Dpt leaf size in absence of DMAR support
Without iommu the DMAR initialization is skipped and the Device page tables
always refer to a single level leaf size. This causes the whole guest
available device memory to be mapped with 4k pages in Hedron, causing OOM
on larger machines. Set the supported leaf level according to the normal
page size support when initializing the bootstrap processor. The DMAR
initialization can still lower this based on detected DMAR capabilities.

Signed-off-by: Stefan Nuernberger <stefan.nuernberger@cyberus-technology.de>