vmx: don't fiddle with IA32_FEATURE_CONTROL when VMX is not supported
On KVM, when VMX is not enabled, Hedron still tries to enable VMX via
IA32_FEATURE_CONTROL. This results in a #GP. Change the code to only
fiddle with the MSR, when the CPU actually says it supports VMX, but
the BIOS has disabled it.

This fixes booting Hedron in Qemu/KVM when Qemu doesn't expose VMX.

Fixes #150.