build: fix dependencies for linker script
I've noticed that we always re-link the hypervisor, even when nothing
changes. The problem seems to have been that our linker script
preprocessing step does not generate a dependency file.

I've added the necessary CMake and gcc parameters to make this happen
and now an unmodified source does not execute the linking step again.