vmx, svm: don't special case EPT/NPT violations
To open a receive window with the right hotspot, the kernel special
cases EPT violations and saves EPT exit information. As we do not
handle EPT violations like page faults, there is no reason for this
obscure semantics and by removing it we have one less special case in
the kernel.