pio: eagerly allocate IO permission bitmaps
Earlier we tried to save 8K of memory by lazily allocating backing
store for the permission bitmaps when they where first modified.

As we rather want to die on PD creation time when we are out of memory
and the complexity of this optimization is not worth the savings, just
allocate everything eagerly instead.

While I was here, I also made sure the permission bitmaps are actually
freed.