doc: rewrite hedron README
We are not doing a good job of setting up developer's for success yet.
Make a pass through the README to fix any obvious issues and also
update it to reality.

One caveat here is that this repository is mirrored to Github and we
don't want to needlessly include links to internal resources. That's
why all internal developer documentation will live in the internal
SuperNOVA wiki.