acpi: refactor fadt parsing
Tuxedo laptops have boot issues due to wrong parsing of the FADT table. To fix
this problem, the parsing has to take the extended address fields in the FADT
table into account.

Refactor the FADT parsing by adding functions that respect the extended address
fields and split the parsing and initializing to make the parsing unit-testable.