acpi: overload the init-functions of the acpi-gas
Tuxedo laptops have boot issues due to wrong parsing of the FADT-table. To fix
this problem and make the parsing unit-testable, the parsing and initializing
has to be split. The added functions will be used in the refactored parsing.

Overload the init function of Acpi_gas to make it more comfortable to use.