kernel page: add the kernel page class
A kernel page is a kernel object that represents a page of hypervisor memory
shared with user space. As this is needed in several cases, this class clears
up existing code paths around creation and destruction of threads/vCPUs and
allows for safely reclaiming shared memory between user space and the kernel.