ioapic: redesign to be self-contained
So far the Ioapic and Gsi classes had an unhealthy relationship. The
Gsi class would keep a cache of IRT entries and use that to accelerate
mask/unmask operations. This meant that from the outside all of the
complexity of IOAPIC programming was exposed.

This commit moves everything as much as possible into the Ioapic
implementation. The Static_array that was previously used for
suspend/resume is now our permanent write-through cache. This has the
added benefit, that there is nothing to save on suspend, because
everything is already stored in memory.