vector_info: leave masking level-triggered interrupts to userspace when reprogramming
We can only partially do the right thing here and it results in
complicated semantics. With the previous approach, we can accidentally
mask level-triggered interrupts when the user reprograms the interrupt
to a different CPU/vector pair. With this change we acknowledge that
the user needs to the right thing here anyway, so we can also leave
the IOAPIC masking to them if they really want to reprogram IOAPIC
level-triggered interrupts.