hip: be extra careful not to write OOB
When we overflow the HIP size, we currently overwrite the boot page
table. This causes hard to debug sadness later and is the reason for
issue #183.

This patch is not a fix yet, but makes debugging significantly easier.
The real fix needs to shrink the HIP or allow for more than one page
of HIP.