ci: drop clang 11 support
The unit tests with ASAN die due to an ASAN bug:
 ERROR: AddressSanitizer failed to allocate 0x0 (0) bytes of SetAlternateSignalStack (error code: 22)

This is fixed in newer clang versions. So don't dwell on the past and
just remove the ancient clang from our CI.

In case, we need actually need to workaround this, we can also set
`ASAN_OPTIONS=use_sigaltstack=false` in the environment.