hip: drop per-CPU LAPIC register dump from HIP
With more CPUs this takes a lot of space. It turns out that most
information is not used in any way. Instead we now only publish to
userspace whether the LAPIC was activated before Hedron started and
how LINT0 is configured. This should be enough for passthrough
virtualization while keeping the HIP small.