sc: fix budget calculation
The time budget of a SC describes the amount of ticks the SC can run
until Hedron reschedules. The budget is calculated from a given value in
microseconds (provided by the user) and the TSC frequency.
Because the freq_tsc value used for the calculation is hold in a 32bit
unsigned, the whole calculation was done in 32bits even the
budget destination variable is 64bit. This leads to the upper 32bit
being cut off if the result is larger than 32bits.
Thus, users trying to set some large time quantum for a SC might have
received a lower quantum.
This commit fixes the calculation so an SC gets the correct time budget
specified by the user.