ioapic: set ID to what the MADT table reports
The MADT contains the locations of the IOAPICs and their IDs. So far
we assumed that the BIOS has programmed the IOAPIC ID correctly, but
this is not the case on some boxes. On a Lenovo SR650v2, we come up
with ID 8 in the MADT and ID 14 in the IOAPIC.

This creates a problem, because we report the IOAPIC's ID in the HIP.
The userspace then uses this ID. Internally, we use the MADT ID in our
data structures. This results in userspace seeing IOAPIC IDs that the
relevant system calls reject.

Fix everything by reprogramming the IOAPIC to the ID specified in the
MADT.

Linux also seems to fix the IOAPIC IDs in the IOAPIC and is not even
that judgemental about it.