irq: make room for more irq_ctrl subops
We need to add another irq_ctrl suboperation to configure LVT thermal
interrupts. Unfortunately, we have already exhausted all bits.

Make space for additional suboperations by claiming bits 11 to 8 for
the suboperation. 16 suboperations should be enough for everyone!

I've dropped the Git hashes in the changelog, because they are easy to
get wrong when rebasing the branch during development. git blame
should do the job just as well for the rare cases where we need it.