kernel page: add a constructor that doesn't need a selector
We want the FPU use a kpage for its xsave-area, but we can't provide a
kpage in the ec_create syscall (due to a lack of free parameters). As a
workaround a KP can be created that the user space has no access to. As
soon as we can provide a kpage in the ec_create syscall, this
constructor has to be removed.