fpu: the xsave area is now a kernel page
This also means that each Ec needs an additional KP for the Fpu. As
there is currently no way to let the user space provide a KP, the EC's
constructor creates a KP for the Fpu. To avoid handling nullptrs in the
Fpu, the idle EC's Fpu also gets a KP.