ld: Demystify magic 0x90909090 constant
In the linker script language, an output section can have a fillexp. This is the
default value that is used to fill padding/gaps. This commit changes the fillexp
to the value we discussed and adds more documentation references.

The fillexp is or best effort approach to fail fast if the IP is at unexpected
addresses.