vmx: add a delete-function to the Msr_area
The Msr_area had no delete-function until now. The newly introduced vCPU
has a unique pointer to a Msr_area to make sure that the Msr_area will
be deleted when the vCPU is deleted. Thus the Msr_area now also needs a
delete-function. While I was here, I deleted the destroy-function, as it
was unused.