vcpu: add missing call to regs.nst_ctrl to the constructor of the vCPU
I thought regs.nst_ctrl had to be part of Vcpu::run, because I thought
that it may use values that are specific to the Ec that is currently
running. After looking at that function for a while I figured that this
is not the case and the function only uses values that are set during
hedrons startup. Thus I added the call to the constructor.