ec: don't destroy GS base
For NMIs we can enter from userspace or from the kernel. In the latter
case, we do not know whether the kernel has already switched to its GS
base (or back). To handle this we add another exception entry path
that saves and restores GS base.

Restoring GS base and KERNEL_GS_BASE is left to the C++ code, if it
needs it.