bootstrap: fix boot race that left CPUs spinning forever
When the BSP leaves wait_for_all_cpus, it re-sets barrier to 0. CPUs
that still haven't seen barrier reach Cpu::online will now be left
stranded and spin forever.

Fix this by moving the rearming to the assembly resume path of the
BSP. At this point, we definitely want to rearm the synchronization
logic.