cpulocal: use rdgsbase instead of using IA32_GS_BASE
We now enable FSGSBASE early during boot. This is nice, because we
don't have to check where we are in the code to use wrgsbase.
Previously, wrgsbase could only be used after Cpu::init is done.

Since we only support machines with FSGSBASE, the missing diagnostic
for really ancient CPUs should be ok to remove.