vcpu: fix GDT limit after VM exit
After a VM exit, the GDT limit is set to 0xFFFF by the CPU. So far we
have never fixed this to the real GDT limit. This allows a user to
reference unintended memory by loading segment registers.

Fix this by making our GDT actually 64K large. This means that the
default set by the CPU actually just works for us and we save roughly
100 cycles of GDT reloading per VM exit.