gdt: use one GDT for the whole system
Instead of having 64K of GDT per CPU, we build one GDT for the whole
system. This means the TSS selector of the current CPU is not a
constant anymore. We also need to fit all TSS descriptors in a single
GDT.

This should be unproblematic until NUM_CPU goes to
8K.