ec: mark spurious NMIs
... with a simple but best-effort mechanism. These spurious NMIs are
meant for injecting into the passthrough guest to enable usecases such
as the hung task or hard lockup detector in Linux.

There is no perfect solution here, because we cannot reliably
differentiate between the passthrough guest or Hedron (or the platform
for that matter) sending the NMI.

The solution we chose here will ignore interrupts that arrive when we
execute in root mode. This is deliberate, as it will automatically
ignore all NMIs that we miscategorize during early boot. If we don't
do this, userspace has to find a convenient time when to start
delivering these NMIs to the guest. If it starts delivering too early,
the guest will crash. The downside is that we have a chance of losing
NMIs.